+++ framework/core/include/pf/net/connection/manager/basic.h
@@ -33,7 +33,7 @@ class PF_API Basic : public Select {
    virtual ~Basic() {};
 
  public:
-   virtual bool heartbeat(uint32_t time = 0);
+   virtual bool heartbeat(uint64_t time = 0);
    virtual void tick();
 
 };
