+++ framework/core/src/basic/tinytimer.cc
@@ -14,15 +14,15 @@ bool TinyTimer::isstart() const {
   return isstart_;
 }
 
-void TinyTimer::set_termtime(uint32_t time) {
+void TinyTimer::set_termtime(uint64_t time) {
   tick_termtime_ = time;
 }
 
-uint32_t TinyTimer::get_termtime() const {
+uint64_t TinyTimer::get_termtime() const {
   return tick_termtime_;
 }
 
-uint32_t TinyTimer::get_last_ticktime() const {
+uint64_t TinyTimer::get_last_ticktime() const {
   return last_ticktime_;
 }
 
@@ -32,22 +32,22 @@ void TinyTimer::cleanup() {
   last_ticktime_ = 0;
 }
 
-void TinyTimer::start(uint32_t term, uint32_t now) {
+void TinyTimer::start(uint64_t term, uint64_t now) {
   isstart_ = true;
   tick_termtime_ = term;
   last_ticktime_ = now;
 }
 
-bool TinyTimer::counting(uint32_t now) {
+bool TinyTimer::counting(uint64_t now) {
   if (!isstart_) return false;
-  uint32_t new_ticktime = now;
-  uint32_t delta = 0;
+  uint64_t new_ticktime = now;
+  uint64_t delta = 0;
   if (new_ticktime > last_ticktime_) {
     delta = new_ticktime - last_ticktime_;
   } else {
     if (new_ticktime + 10000 < last_ticktime_) {
       delta = 
-        (static_cast<uint32_t>(0xFFFFFFFF) - last_ticktime_) + new_ticktime;
+        (static_cast<uint64_t>(0xFFFFFFFF) - last_ticktime_) + new_ticktime;
     } else {
       return false;
     }
